+++ composer.json
@@ -1,7 +1,7 @@
 {
-    "name": "react/chatroulette",
-    "description": "Chatroulette sample application.",
-    "keywords": ["chatroulette"],
+    "name": "react/ExampleChatWithRandomStranger",
+    "description": "ExampleChatWithRandomStranger sample application.",
+    "keywords": ["chat"],
     "license": "MIT",
     "require": {
         "php": ">=5.4.0",
@@ -12,7 +12,7 @@
         "phpunit/phpunit": "3.7.*"
     },
     "autoload": {
-        "psr-0": { "React\\Chatroulette": "src/" }
+        "psr-0": { "React\\ExampleChatWithRandomStranger": "src/" }
     },
     "config": {
         "bin-dir": "bin/"
