+++ phpunit.xml.dist
@@ -5,7 +5,7 @@
          bootstrap="tests/bootstrap.php"
 >
     <testsuites>
-        <testsuite name="Chatroulette Test Suite">
+        <testsuite name="ExampleChatWithRandomStranger Test Suite">
             <directory>./tests/React/</directory>
         </testsuite>
     </testsuites>
