+++ tests/React/ExampleChatWithRandomStranger/LoggingAppTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace React\Chatroulette;
+namespace React\ExampleChatWithRandomStranger;
 
 use Monolog\Logger;
 use Monolog\Handler\TestHandler;
@@ -16,10 +16,10 @@ class LoggingAppTest extends \PHPUnit_Framework_TestCase
     {
         $this->handler = new TestHandler();
 
-        $this->logger = new Logger('chatroulette');
+        $this->logger = new Logger('ExampleChatWithRandomStranger');
         $this->logger->pushHandler($this->handler);
 
-        $this->appMock = $this->getMock('React\Chatroulette\AppInterface');
+        $this->appMock = $this->getMock('React\ExampleChatWithRandomStranger\AppInterface');
 
         $this->app = new LoggingApp(
             $this->appMock,
